Large Update, Mostly Items Tab but other minor changes around too.
Created several new files for the Item Sections (Base and Combined, 'BS, 'CS'). These cells both contain collection views within them that are used for displaying different items. Base section is a compositional layout that has a scrollable collection view at the top where a users presses the base item they are interested in, the collection view below will then reload with items that are formed from the selected base item.
The combined section is currently unfinished but currently has a scrollable collection view of all the combined items in the game. This will eventually be selectable so that when a user taps one it will load and display a view of that items stats, information and description.
Deleted ItemPageCell file.
Changed MenuBarController back to using indexPath.section instead of .item
Created a StatType Enum that is used when creating Item Model to convert the parsed firebase data keys to a more understandable word.
Added more ReuseIds to the Constants enum, but these will eventually all be removed once other cell conform to Resuable Cell Protocol.
Create UICollectionView Extension that has a generic custom dequeuReusableCell function, will use this throughout the app from now on and will refactor old code to utilise this. Base Cell also slowly transition to using a ReusableCell Protocol to help structure and maintain code organisation.
Created a StatView used within the ItemsTab and Champion Cells, contains a switch statement that takes a StatType to handle which image and image tint color to show within the Stat View.