Custom Team Comps Commit!
This commit includes a lot of changes.
Began to implement the functionality where users can create their own custom team compositions and save them to the app.
This commit includes a lot of code reorganisation and new code added. Currently the functionality allows for users to create and save team comps but they can't view the detailed version of them. Committing now to have a nice roll back spot as the amount of uncommitted code was growing quite large.